:rocket: Use deployments to scale application PODs
PODs alone cannot help scale out application, so we use k8s deployments
to encapsulate such PODs and create a deployment which makes it easier
to manage all resources.
We can also create replica sets, but the advantage here with
deployments is that we can get rolling updates into our application
which can be tracked using the update history and easily be undone.